support normalizing dependencies
Summary:
When we dump shared memory to SQLite, we no longer understand the underlying representation of sets. In order to support the comparison, we need to either

a) Change the representation of dependencies to be stable (i.e. sorted list). This comes at a time complexity cost for this case, as we'd need to support linear insertions.

b) Change the shared memory representation just before dumping to SQlite, and remove the old, non-canonical keys since they might cause inconsistencies. I disliked the destructive nature of this approach.

c) Somewhat break the abstraction barrier by normalizing how we construct the sets (make sure they get constructed from the same function with the same list passed in as an argument).

c) seemed like the most appealing option to me.

Reviewed By: dkgi

Differential Revision: D14248781

fbshipit-source-id: a649936f79b34d11dd49c04116c7d5961585da7b