fix how we analyze files changed multiple times
Summary:
The Deferred structure had a bunch of complicated logic to add the concept of a visited deferred request, put in place so that we don't need to re-analyze files multiple times if recursive deferred requests ended up adding the same dependency. However, we had a bug in which we didn't clear this correctly, so a file would only get analyzed once.

Previous changes have made the dependency computation more efficient and therefore obviated the need to have a recursive deferred request structure, so simplify it accordingly, eliminating the bug as we go.

Reviewed By: dkgi

Differential Revision: D14269110

fbshipit-source-id: e151df3fa7b211b42bdf540d803f171d60a0371e