Fix inconsistent override when comparing type variables
Summary:
Ran into this with the example:

```
_T = TypeVar(bound: int)

class Parent:
   def foo(int) -> int: ...
class Child(Parent):
   def foo(_T) -> _T: ...
```

This should be fine, but instead errors:
`Inconsistent override [14]: Child.foo overrides method defined in Parent inconsistently. Parameter of type Variable[_T (bound to int)] is not a supertype of the overridden parameter int.`

We're OK with comparing (using `TypeOrder.less_or_equal`)
Type.Variable <= some_type
but not
some_type <= Type.Variable (always false)

When comparing parameters, we want to know if `int <= _T` in this case. This returns false even though this override is fine. To cover this and more complex cases, I think we can just check whether it's possible to instantiate `_T` to `int`, or whichever other type we're comparing to. If so, then the override is alright.

Reviewed By: dkgi

Differential Revision: D14238505

fbshipit-source-id: 3d33ab35958ded315a3730bdd233e21c86c91448