Option to run upgrade without local configs
Summary:
Motivation - users in distillery would like to be able to run upgrade and have it take care of linting. This couldn't happen before because we relied on piping pyre output in distillery directly into upgrade, and couldn't invoke the script independently (it relied on local configurations).

This supports `pyre-upgrade fixme --run --lint` in distillery which will run both pyre and lint.

Reviewed By: sinancepel

Differential Revision: D14304223

fbshipit-source-id: de7976da6dde6c80a1a406be52e6f1ff53216fb8