considering overloads in solve and less or equal
Summary: This was pretty straightforward, unfortunately this generated a bunch of errors because Dict[bottom, List[bottom]] isnt leq Dict[int, List[int]], even though Dict[bottom, bottom] is leq Dict[int, int] and List[bottom] is leq List[int].  Problem is that the special casing for bottom only goes one layer deep.  This is relevant to the quest for D14243734.  Seems like we need to get rid of this behavior of bottoming free type variables and instead allow them to propagate with a kind of simulated bind just like simulated signature select, until hitting a assignment or a determine, at which point we require an explicit annotation, which we can solve against, and bottom if not supplied.

Reviewed By: dkgi

Differential Revision: D14263986

fbshipit-source-id: d16c7cbabc776f13622d2bbb63f852faa0928732