Refactor attributes to use identifier name
Summary:
I don't understand why the name of an attribute (that contains a field for parent, so qualification isn't the issue here) can ever be a chain.

It wouldn't make sense to retrieve the name of an Attribute (or a Class or a Method for that matter, but those diffs can come after) and receive an access that contains attribute accesses or calls.

Need to clean up the overuse of Access.t before deprecating it in favor of Attributes and Calls.

Reviewed By: dkgi

Differential Revision: D14330598

fbshipit-source-id: e266a5be3c1f12db310bd536b4623bc2d4ddd9e1