Properly preprocess concatenated format strings
Summary:
Our goal is to properly parse string literals that are:
- Concatenated by juxtaposition
- Both plain string literals and format string literals are involved.

Here are some examples:
```
  x, y = 'A', 'B'
  s0 = '{x}' '{y}'    # s0 = '{x}{y}'
  s1 = f'{x}' '{y}'   # s1 = 'A{y}'
  s2 = '{x}' f'{y}'   # s2 = '{x}B'
  s3 = f'{x}' f'{y}'  # s3 = 'AB'
```
Currently we handle this by:
1) Having the parser read all the literals in and concatenate them (i.e. `s0`, `s1`, `s2`, `s3` will all gets parsed into a `String` with value `'{x}{y}'`.
2) Having a preprocessing step to extract expressions from the concatenated value by manually parsing strings enclosed by curly braces.

The current solution is not ideal because the preprocessing step can only see the concatenated string -- it does not know how many pieces the concatenated string consists of. More importantly, it does not know whether each piece is a format string literal or a plain string literal. In other words, the current approach does not parse `s1` and `s3` correctly. Such a behavior would cause troubles when typing checking a program like this:
```
x = 'A'
print(f'{x}' '{y}')
```
Pyre would incorrectly complain that the global variable `y` is undefined because it does not distinguish the status of `f'{x}'` and `{y}` --- the former is a format string from which expressions can be extracted, while the latter is just a dumb string literal.

The solution is conceptually straightforward: we just need to expose the knowledge of plain/format string literal from the parser to the pre-processor. Concretely, a new structure `StringLiteral.Substring` is introduced to the AST, and the parser and the preprocessor are taught to utilize it.

Note that f-string can be nested, and properly addressing the nesting is not part of this diff. Previously we did not handle them anyway and this diff will not make things worse.

Reviewed By: shannonzhu

Differential Revision: D14309304

fbshipit-source-id: bd63d90c468ddc8cfd4d9fd237693370966b51fb