Clean up unqualified name for defines
Summary:
`def x.y()` is a syntax error - you can only create defines with identifier names.

The reason we keep these as Accesses is because we need to keep track of qualification. This generates some ugly hacks all over the place (esp. see statement.ml for examples) - it would be cleaner if we can do the exception throwing in one place in the case where the Define name does not end in an identifier.

Reviewed By: dkgi

Differential Revision: D14344318

fbshipit-source-id: d6194d992be75dca2b783e0562aa6bd19d263117