only normalize targets once when building
Summary:
The old implementation of `generate_source_directories` called itself recursively in order to pick up on newly generated source directories. This isn't ideal, since it has an expensive normalize call into buck, and we'd rather only have 1 call in total.

Move the commonly needed functionality out into a private helper, and re-use the normalized targets. The build flag stays because it's used to "always build when needed even if not prompting". This would probably be a better "build = | AlwaysBuild | Prompt | NeverBuild", but that's a different refactor.

Reviewed By: shannonzhu

Differential Revision: D14352321

fbshipit-source-id: 13fbc0bd983cf09af34e25d3591b7d7c3766c5c5