stop ignoring typing.Hashable
Summary:
This has perf implications if we're doing conformity based on inherited methods (which would be correct) but because we're currently not doing this, this has no major perf hit.
The alternative would be to put the special case that everything is hashable into the type order.
As long as this doesn't have perf costs, this is simpler.

Reviewed By: dkgi

Differential Revision: D14370873

fbshipit-source-id: 715c3b3effdeb9dec6c42a313aaed6ed38571500