Qualify decorators in global context
Summary:
Method defined inside a class were not correctly qualified previously. For example,
```
from somewhere import decorator

class A:
  decorator
  def foo(): ...
```
In the above code, `decorator` will not be expanded into `somewhere.decorator` by the preprocessor. This behavior would cause Pyre to emit "Global name not defined" warning if decorators were type checked (which will come in a followup diff). Curiously, if `foo()` is defined outside of `A`, the qualification works fine.

This diff fixes the issue by forcing the qualification procedure to always look for forward reference.

Reviewed By: dkgi

Differential Revision: D14369883

fbshipit-source-id: d8a972788f4a4fdcd247dd24a19422afc616cddc