Fix incompatible var false positives in enums
Summary:
Got several reports that this is borked this week :(

We used to error on assignment in enums, since the "expected" type is an instance of that enum but the type the checker gets is an int or whatever the base type is. Since the base type is not < the enum, the typechecker complains.

Fix this to only complain when an explicit annotation is given and that annotation is bad.

Reviewed By: dkgi

Differential Revision: D14378208

fbshipit-source-id: 1c710b9e769bbbb39299f30dcf7767e1b6832dba