support extracting a little endian representation of a MD5 digest
Summary:
Values in the shared heap are stored as (hash(key), serialized_value) tuples, so we have no way of getting keys directly from a SQLite dump of the pyre server's environment. Therefore, we'll need to either change hack_parallel (probably not a good idea) or construct a mapping from hash -> actual key value, which is the approach I plan on going with here. This primitive will be used to support a pyre query to get a mapping of all (hash -> key values) for a server (which we can do since we store most keys needed for type checking), which we'll use to give meaning to inconsistencies.

An alternative implementation would be to use C FFI to re-use the hack_parallel code directly. However, it seemed a bit less tasteful to expose the library's function from this detail.

Reviewed By: dkgi

Differential Revision: D14332210

fbshipit-source-id: cfc468041b553a3a285cf3ad20c84cdcf4b35104