Add migrated_id field to TraceFrame
Summary:
Add a migrated_id column to the trace_frames table. This is a temporary field
for migrating existing pre/postconditions to trace frames.

Since there is nothing unique that identifies each pre/postcondition aside from
the ID, we need to track the corresponding trace frame for each condition that
has been migrated.

This mapping helps when populating the issue_instance_trace_frame_assoc from
issue_instance_*condition_assoc (and trace_frame_message_assoc from
pre/postcondition_source/sink_assoc), especially in the case where the
migration script fails/aborts due to unexpected failures.

Basically, I need the migration script to be able to restart from the state it
was last in when it failed, and this column tracks the state. Column will be
dropped once migration is complete.

Reviewed By: fahndrich

Differential Revision: D14309549

fbshipit-source-id: 1a70f46ed1e3022405fc624cf3e538b217dc9d36