supporting callback protocols and laying the groundwork for lazy protocols
Summary:
mypy supports callables implementing protocols that only have a __call__ method (https://mypy.readthedocs.io/en/latest/protocols.html#callback-protocols).

This is important for taking complex callback types, as the Callable syntax gets really awful really fast.
This protocol inference is impossible to do ahead of time as we don't know the set of all possible callable types in a program, which means we have to do this lazily.
As we are planning to transition all protocol comparison to be lazy, this was a good opportunity to lay down the necessary plumbing to do all protocol comparisons here, hence why the interface was created to accept all types and not just callables.

Reviewed By: dkgi

Differential Revision: D14377474

fbshipit-source-id: fd171d9cf743c00e8c319550d4870a794c0f6451