Change ProjectFilesMonitor to throw exceptions in constructor
Summary:
Prior to this we were waiting until the monitor was daemonized to check whether we could actually find a watchman root. It would be better to raise an exception during construction rather than fail silently.

Note that this class isn't being used yet, so no call-sites need to be changed.

Reviewed By: dkgi

Differential Revision: D14379975

fbshipit-source-id: fd1c3e65e4c36e98a5ef5529b85903f2f2360fe8