Suppress errors on special methods
Summary:
whereicometocry

There are so many of these ignores in distillery that I think it's starting to become better to take the false negatives over the false positives. This will be resolved when special methods are supported, which is blocked on our access refactor (see T35601774 and D14315713 for more explanation).

We're interested in suppressing errors of the form `Type[Foo].__special_method__()`, because this results in us calling the method `Foo.__special_method__(self, params)` rather than `Type.__special_method__(params)` where `self` is already bound to `Foo`.

* Note: the above behavior is correct when `Type[Foo].__special_method__()` was the actual call, but often this call is preprocessed from `special_method(Type[Foo])` in which case we need the other behavior.

This happens because `x: Type[Foo] ` followed by `x.__special_method__()` is indistinguishable from a direct access-as-class-method: `Foo.__special_method__()`.

Unfortunately this can't accurately suppress errors, because the choice between an instance method of the meta type and a class/static method of the type occurs before the signature is checked and the error is generated. The error contains no information on which of the above was chosen (and we also don't know at error generation, so we can't add logic to not throw the error to begin with). In either case, all we see is that the base of the method is `Foo`; due to the structure of access fold, all we have at each step is the resolved of the previous steps and the next identifier.

Since these false positives all manifest as arity errors on special methods where there is parent, I figured we won't likely be suppressing too many true positives in order to get rid of a lot of false positives (ex. the ones pointed out by dominik in the last upgrade diff: D14405832)

Reviewed By: dkgi

Differential Revision: D14413641

fbshipit-source-id: 38653c7fedcb3d1edd18032a5ec994e20e0cde8d