simplify watchman subscriber names
Summary: We were using the class name of the subscriber for the lock/log/pid files (e.g. ProjectFilesMonitor -> `.pyre/projectfilesmonitor/projectfilesmonitor.pid`). We also had a separate `_name` property used in logging (kept separate to keep logging consistent with old logs). This is unnecessary complexity, so we get rid of using the class name and just allow the child class to specify what name to use for its files.

Reviewed By: sinancepel

Differential Revision: D14423722

fbshipit-source-id: 69ce414fce0643e3d4c6e6602d77f95cdb6d81f2