Turn on python watchman daemon
Summary:
Spawn watchman daemon when the server starts.

To simplify rollback, we hard-code `use-watchman` to be `False` so that the OCaml daemon is not spawned, so on rollback we need only to revert the client code. T41488848 was created to track the eventual removal of the OCaml daemon (and cleanup of this hard-coding)

Reviewed By: dkgi

Differential Revision: D14417545

fbshipit-source-id: 57f1e6ec1554173c7417486ac61414fd47611a69