convert all shared memory to be decodable
Summary:
The basic changes are pretty simple:

Move the contents of the decodable module into memory, since that's the module all shared memory goes through.

Change the inclusion of HackParallel.SharedMem to resolve module naming conflicts.

Now that we require keys to be decodable, augment keys as needed.

Most of the diff deals with the fallout.

Reviewed By: dkgi

Differential Revision: D14451296

fbshipit-source-id: 5264eefa8c26e7410b6e21d41bde11220239bbf5