Allow for use of generic types in function bodies
Summary:
In order to directly construct generic types like `Something[int]()` we need to implement a version of `Generic.__getitem__` that has parameters corresponding to each of the generic parameters, and returns a Type[...] corresponding to those parameters.
In general terms we want `Generic[ChildClass, T1 ... Tn].__getitem__` to have type `(Type[T1], ... Type[Tn]) -> Type[ChildClass[T1, ... Tn]]`
This gives us both enforcement of the bounds on the variables and the correct number of variables.
This is impossible to do within the stubs, even with variadic generics, as you can't subscript type variables as I did in my imaginary prototype above.

Reviewed By: sinancepel

Differential Revision: D14452232

fbshipit-source-id: 65e9e81d97abc78de2d43c9bf9c33e6f0de902dc