Fix async iterators/generators and callables on classes
Summary:
D13864544 introduced better support for typing async functions and
async generators. The newly introduced check relies on checking the body of a
method for yield statements.

For implicit attributes the `body` field was not stored, making the check
impossible. Keeping the body statements around allows the checks to run
successfully.

Reviewed By: shannonzhu

Differential Revision: D14475128

fbshipit-source-id: ff55a3791efeb8de94dc8e5a63d2efce1b40056f