Move symbolic link management to SharedAnalysisDirectory
Summary:
Rather than have the buck module or the file monitor be aware of how to manage symbolic links, it makes sense to have the `SharedAnalysisDirectory` handle this information. This also avoids creating a symlink mapping for the case of a simple `AnalysisDirectory` (which has no symlinks).

The 2 main things this change introduces:
- Move symlink creation/deletion/tracking to `SharedAnalysisDirectory`
- Add `update` method to `AnalysisDirectory` and `SharedAnalysisDirectory` which takes a list of updated paths and returns the paths Pyre cares about (for now, just return the paths which fall under the directory root or search paths)

Reviewed By: sinancepel

Differential Revision: D14464130

fbshipit-source-id: a05660152ee55916887f091de7551aa2c3849820