Generate new access from identifiers
Summary:
This is at toplevel expression to avoid creating more record types. :/

Also, this new access form supports locations at each element rather than a single location at the head of the chain plus locations on call elements only. Will be mindful of the perf impact when we start migrating over... if it turns out to be terrible I'll cut that.

Reviewed By: sinancepel

Differential Revision: D14468051

fbshipit-source-id: 1c2689375fbed8dfd4c9438229efd70d871ac876