decode all of environmentSharedMemory properly
Summary:
Now that we have a good enough (tm) hashes -> keys mapping, we should decode as much of shared memory as possible.

The tests here aren't comprehensive as they were repetitive to write and mostly check that I put in the right module. Happy to write one for each kind if desired, but seemed like it'd just complicate the review rather than adding much value.

Reviewed By: dkgi

Differential Revision: D14483317

fbshipit-source-id: f4076d0eeb79726140ee11f5a6db4615c4a6405d