Perform type check on decorators
Summary:
This diff teaches the type checker to look at decorators on `Define`s and perform checks on them along with the body of those `Define`s. The immediate benefit is that if the decorators are returned from a callable, we get more coverage by checking the types of the arguments.

The essential part of this diff is just the 7-line `check_decorators` function. The rest of the changes are all about suppressing false alarms and fixing broken tests.

- Since decorators are not checked before, existing tests used them quite liberally without paying too much attention to whether the decorator names are imported. Several class/method stubs need to be added to make sure Pyre does not complain about missing definitions.
- One big source of false alarms come from decorators that has imprecise annotations (see T41382604). We whitelist all decorators that do this in `Recognize.whitelisted_decorators` and defer the task of fixing them to the future.
- Another big source of false alarms come from decorators that produce other decorators (see T41383196). We currently white list known decorators that are derivatives directly in `check_decorators`.

Reviewed By: dkgi

Differential Revision: D14327806

fbshipit-source-id: 16f1b1a833662f14b4cc169e9e525d621f21abca