Don't set deps regardless of sandcastle in upgrade
Summary:
We keep needing to manually unset the deps on automatically submitted diffs; upgrades on projects might happen in batches but don't rely on each other.

The sandcastle command invokes pyre upgrade and the sandcastle field is not actually set. Didn't think it was worth distinguishing to only do this in sandcastle, seems fine for all auto-submitted diffs.

Reviewed By: dkgi

Differential Revision: D14536994

fbshipit-source-id: bd57a3585760ca6c826bbc11517f646e987fe84d