Fixing is_concrete
Summary:
is_concrete was only checking for bottoms when they were at the top, presumably because the original implementor didn't want to trip on None = Optional Bottom, which can't be excluded with exists.  The full Transformer interface handles this easily.

This avoids suggesting someone type something as List[], which would be an error.  This came up in an error report on its own and also is important for not inferring generic constructors.

Reviewed By: sinancepel

Differential Revision: D14523850

fbshipit-source-id: d94971df7426b55f8e63686352f365e22b8f5f24