Compute symbolic links for SharedAnalysisDirectory after prepare
Summary: We were building our symbolic link mapping when the object was created. Often the analysis directory would still exist from previous runs, but when the analysis directory did not (e.g. on a `pyre check`) this step would fail with a CalledProcessError. We should instead defer this computation until after the directory is built.

Reviewed By: dkgi

Differential Revision: D14540906

fbshipit-source-id: dd084ab61a39160a344c13281be8daf62480f6c3