Fill missing and mis-filled generics with Anys
Summary:
PEP 484 specifies that
> If a generic type is used without specifying type parameters, they assumed to be Any

We have that handled for inheritance, but doing so for annotations turned out to be trickier.
In general the approach was to add correction behavior to `Resolution.parse_annotation` toggled by a `correct_parameters` labeled argument, and adding errors in `parse_and_check_annotation`

There were some interesting wrinkles to this approach.
1. There were many places where we just wanted primitive types, and weren't really annotations being parsed per se.  This is arguably an abuse of the interface and should have another function
2.  In order to preserve the principle of local strict-ness, we need to only error on completely missing parameter types in the file where they are introduced.  This means correcting aliases, which was easiest to do by only passing the alias map into resolution instead of the whole parse_annotation.

Reviewed By: dkgi

Differential Revision: D14446964

fbshipit-source-id: 951f9849d394accd0ec130294abe2fd8a240ce84