replacing mismatch with any with is_consistent_with
Summary:
The existing implementation of `mismatch_with_any` was incomplete in a number of ways, and for structural reasons was impossible to fix.

We considered:
* trying to replace covariant anys with Top and contravariants with Bottom
* implementing a bidirectional solve
* introducing types who generally act differently covariantly and contravariantly

Implementing is_consistent_with was relatively straightforward, and allows us to get the behavior we want, reusing the machinery already built for less or equal.
This was decided to be the lesser evil.

Reviewed By: sinancepel

Differential Revision: D14485054

fbshipit-source-id: 6629e698ef6c17c6f1fe2a0b923d9b4678c253f5