fix the dependency registration for builtins
Summary:
Relatively minor bug found by the consistency checker: We were registering the import dependencies for the raw name of module being
imported instead of qualifier. This works in most cases, but doesn't work for builtins.

The reason we're registering dependents for the empty qualifier here is that we rely on builtins.pyi = empty qualifier throughout the codebase, and that undertaking would be much more significant than this change, which'll trigger type checks when the builtins stub gets changed.

Reviewed By: dkgi

Differential Revision: D14529348

fbshipit-source-id: aa36ae8c43c09708810f00fa3101e733053b7daf