rely on DependentKeys to get a full set of dependents to normalize
Summary:
We were assuming that the only modules that needed to be normalized corresponded to actual handles. This is not the case, as we can do
`from nonexistent import *`, in which case the dependent still gets registered. Whether it makes sense to register this dependency is questionable, but make the normalization take these into account by getting keys first.

Reviewed By: dkgi

Differential Revision: D14533557

fbshipit-source-id: a619b4d147ef13f570837cfb60890abcd9a2031e