Make identifier the base of reference
Summary:
Thought about this for a while... original concern here was trying to separate the concepts of AST objects from pointer-like references that only exist in metadata of real AST objects.

In the end I don't believe it makes any sense to have two versions of "string", one for AST objects and one for references -- Identifiers are building blocks, but are not AST objects any more than integers are. Creating a node out of something and adding on locations, other AST nodes, etc., turn the building blocks into AST objects. I think it's fine that we have one wrapper around the string primitive, that can be used to create AST objects or references.

There's also quite a lot of value in permitting the single-unit of Reference to be specified as the type of some metadata; often times we know for sure that there is only one string, not a list (see `AnnotatedClass.mli` for a whole slew of these that would have to generalize to `Reference.t` otherwise, which is a lot of unnecessary overhead and information loss). Exposing a bare `string` rather than `Identifier.t` (or creating a parallel wrapper to Identifier) is messy or overly complex & confusing, respectively.

Reviewed By: grievejia

Differential Revision: D14580144

fbshipit-source-id: a8eac951b6e90350c13f3ccac7e58cdf3defae85