Only refine the type of `x` for `x in Iterable[t]` when appropriate
Summary:
If `x in Iterable[t]` succeeds, we infer `x` <: `t`: D14360817

But this should only take effect when `x` is a local variable with inferred types. For global variables or locals with annotated types, let's bail out and try not to be too smart. See linked task for an example where being too smart lead to sadness.

Reviewed By: sinancepel

Differential Revision: D14625465

fbshipit-source-id: 9e20b25b8bbf8a318f9b3edae27c7879d5debf2d