Fix filter_root for SharedAnalysisDirectory
Summary:
SharedAnalysisDirectory was using the base AnalysisDirectory implementation of get_filter_root, which would use the analysis directory root as a fallback if no filter paths were provided (which happens when no targets or source directories are passed on command line and we are not using a local configuration). The result of this is that we would use the shared analysis directory (in the scratch path) as the filter directory, and then all actual files would be filtered from typechecking.

This change uses the location of the `.pyre_configuration` file as the fallback.

Reviewed By: dkgi

Differential Revision: D14542506

fbshipit-source-id: 1cf37d7cf80014457b9e7da23f3711972e61cae7