Make it possible to synchronize on file monitor initialization
Summary: Currently we spawn an anonymous file monitor after starting the server and accept that it be initialized eventually. For the sake of testing, it is helpful to retain the file monitor and have a means for the monitor to communicate when it's finished initializing (i.e. watchman has sent it the initial message).

Reviewed By: sinancepel

Differential Revision: D14623168

fbshipit-source-id: 63a145a0216c3cb6ec78f2763fcbd73dd57b9dee