Preprocess the first argument of `typing.cast` from string to type annotation
Summary:
The first argument to `typing.cast` should be a type object. But it can also be a string. This diff preprocess the latter into the former.

Note that we have to do this before the qualification pass in order to qualify the parsed annotation. As a result, we cannot resolve the qualifier of `cast` and need to hardcode some cases there. Hopefully this is ok in practice.

Also note that we still can't process annotations like `List["Foo"]`. We could alway encourage people to write `"List[Foo]"` instead, though.

Reviewed By: dkgi

Differential Revision: D14589709

fbshipit-source-id: 2da407e0c94eac8454f3c30d149e1d4ff7f6872d