Migrate performance log from plain text to structured JSON
Summary:
As proposed in D14533162, for local logs we want to move the performance-related stats away from plain texts into structural representation. For scuba-oriented data, we want to leave things are they are for now.

This diff kills the `Performance` log section and attempts to redirect the info towards a dedicated `Profiling` module. Internally the `Profiling` module will look for `Configuration.profiling_output` and log to this file if it is specified via a command line argument. Profiling events are logged as a list of JSON strings separated by newlines (which introduces the additional constraint that event names must not contain newlines).

Reviewed By: dkgi

Differential Revision: D14610183

fbshipit-source-id: 28de22cfc064c0281925c304dab38c667a14d8ec