allow added stubs to override python files in the server
Summary:
Context: We had previously run into issues where editing "a.py" could override the stub definition for "a.pyi", which would cause us to clobber the definitions in the stub. To fix this (old) bug, we added a check - if you're updating a file, make sure that Pyre doesn't override a module that already exists.

However, this led to a new, more subtle bug: When you introduce a new stub file for a python file, we don't pick up on that on the server. Change the behavior of the server to support overriding in that fashion.

The integration test with both "a.py" and "a.pyi" being present hopefully presents the issue clearly.

Reviewed By: grievejia

Differential Revision: D14628039

fbshipit-source-id: 7ddf6dd8d426057f2825e68625aa40dd3769ad49