Add a message to incremental requests delineating the end of parsing
Summary:
The way our client display works is that it updates on a Log.info message. For incremental requests,
we weren't writing a message when starting to update the environment. Make it so that we print a message to avoid it looking like the Pyre server is stuck on parsing.

Reviewed By: dkgi

Differential Revision: D14631926

fbshipit-source-id: db8cce55435862f1b884d138d19fca477109f191