Centralize the job of removing/re-adding values when environment building in the server
Summary:
A pain point in updating the environment of the server is that in cases where we want to update the value pertaining to a key, we'd have to "remove_batch singleton(key); add key new_value" all over the place. This is less than ideal as it means that our code has to be conscious that it's removing things, and more importantly this can only run on the master process.

The hack shared memory library provides a solution for this, called the LocalChanges module. This module essentially attaches with each shared memory module a hash table from key -> action, that can be manipulated in local memory without being committed to shared memory. The advantage of this is that we can do many iterative changes to a single key without having to worry about the garbage we're creating when we remove and re-add a value, and architecturally allows us to not have to worry about removals when we're building the environment.

This will be particularly nice when we consider parallelization; the existence of `remove_batch` in populate makes it impossible to parallelize environment building. This design will let us do it per worker.

Reviewed By: grievejia

Differential Revision: D14634886

fbshipit-source-id: f8d5c0c553b8475f1ff876ab7b64f39fe4f2a6da