only use shared memory to populate the environment
Summary:
We were using an in-process collection of hash tables to populate the environment as it was faster single-threaded compared to adding/reading from shared memory. With the LocalChanges module, we can now handle local changes for single processes. Doing so allows us to add everything to shared memory without incurring a large overhead.

Also thread in the scheduler to populate now that we can parallelize environment building.

Reviewed By: dkgi

Differential Revision: D14677184

fbshipit-source-id: ac4ef2c4b44c2c02216096256c1bec24f190deb4