Add db version to run
Summary:
I want to make it easier to write db migrations scripts.

E.g., I need to migrate data where I fill in shared texts for the new id fields, but it's difficult without extra metadata.

The idea is that we can bump the current version when we introduce db changes and then see which runs were written with the new stuff, and which runs need to be upgraded.

Reviewed By: zlandau

Differential Revision: D14656141

fbshipit-source-id: 2f13fc3d7bcdcd6385f15cd29c0a7ddf14ea38be