Add helper Profiling function to make it easy to track duration events without going through `Statistics.performance`
Summary:
`Statistics.performance` requires `Timer` setup and may log to scuba, which can be inconvenient at times if we just want to do some quick local profiling tests.

This diff adds `Profiling.track_duration_event`. The idea is that whenever we want to track how long `do_something` takes in the following code:
```
let x =
  do_something y z
in
...
```
We can simply do this:
```
let x =
  Profiling.track_duration_event "time_to_do_something" ~f:(fun _ ->
    do_something y z
  )
in
...
```

Modifications in the test file look involved at first glance, but it's really just factoring out a piece of logic that was once private in `test_event_format`, so that the newly added `test_event_track` is able to use it.

Reviewed By: MaggieMoss

Differential Revision: D14716131

fbshipit-source-id: 1fefeba3b1e5e569a29a251b0b83931ce00a7972