don't attempt to register literal alias `module.*`
Summary:
Our alias registration viewed the form `from x import *` as literally attempting to construct the alias `module.* -> x.*`, which succeeded in some edge cases. Don't register this literal.

Note: We were not registering each individual component of the starred import as a behavior beforehand either, so this is not a regression. If we'd like to support that, that's a separate feature request.

Reviewed By: dkgi

Differential Revision: D14708708

fbshipit-source-id: 7c76ccf6abeb53349d076f1b46aaaea47cc46123