Only stop pyre with monitor if a related local configuration changes
Summary: Right now the monitor stops pyre when it receives updates to any `.pyre_configuration.local` files. So, even if the local configuration relevant to your own project does not change, a change in a project under the same watchman root will stop your server. This adds a check that the changed file belongs to your project.

Reviewed By: sinancepel

Differential Revision: D14731644

fbshipit-source-id: b51ab0eaeb6e11b68314e031b641df5617ab5d79