Don't throw missing annotation errors on typed params used as attrs
Summary:
We allow the following:

```
def __init__(self, a: str) -> None:
   self.a = a
```

to go without a type annotation on the attribute assignment since the explicit annotation was given in the param list. There are a lot of constructors with long lists of typed params that avoid a tedious experience this way.

Overlooked the fact that if `a` has a type containing Any, our missing annotation logic didn't know to skip these cases (even though this type transfer is already supported in attribute lookup).

Reviewed By: grievejia

Differential Revision: D14721113

fbshipit-source-id: fa9e845f4c1f4c5c6b90118caf7981f43b778c3a