Parse Buck TARGETS files into python objects
Summary:
This is a first step, showing a general approach for parsing Buck TARGETS files into python objects that we can then manipulate on a higher level.
We use the `ast` module to generate a python AST out of a TARGETS file, and then check if each top-level expression is a call with an identifier corresponding to a build rule we support (e.g. `python_binary`). If so, we inspect the AST for the fields we need to build the target and its dependencies.

Reviewed By: dkgi

Differential Revision: D14688772

fbshipit-source-id: bf48293695ed0f61d2a15cda4e6d34005a8c56a9