Add generalized support for parsing non-python buck rules
Summary:
There are some buck rules which we don't care about building, but we still want to parse them so that all the dependencies for a target can be resolved. This introduces a general way to create new rules, using cpp_python_extension as the first use-case.

cpp_python_extension normally compiles C into a .so file which can interface with python through the C-API. Since pyre does not analyze such files, we can treat it as a no-op for our purposes.

Reviewed By: dkgi

Differential Revision: D14725499

fbshipit-source-id: 4864f04f4dfd74e3906d5f749e5bb09aeba05823